Refactor locations and names of controllers before adding more for QR code labels
